+++ apptentive/src/main/java/com/apptentive/android/sdk/storage/Device.java
@@ -36,7 +36,6 @@ public class Device implements Saveable, DataChangedListener {
 	private String localeLanguageCode;
 	private String localeRaw;
 	private String utcOffset;
-	private String advertiserId;
 	private IntegrationConfig integrationConfig;
 
 	private transient DataChangedListener listener;
@@ -102,7 +101,6 @@ public class Device implements Saveable, DataChangedListener {
 		clone.localeLanguageCode = localeLanguageCode;
 		clone.localeRaw = localeRaw;
 		clone.utcOffset = utcOffset;
-		clone.advertiserId = advertiserId;
 		if (integrationConfig != null) {
 			clone.integrationConfig = integrationConfig.clone();
 		}
@@ -372,17 +370,6 @@ public class Device implements Saveable, DataChangedListener {
 	public void setUtcOffset(String utcOffset) {
 		if (!StringUtils.equal(this.utcOffset, utcOffset)) {
 			this.utcOffset = utcOffset;
-			notifyDataChanged();
-		}
-	}
-
-	public String getAdvertiserId() {
-		return advertiserId;
-	}
-
-	public void setAdvertiserId(String advertiserId) {
-		if (!StringUtils.equal(this.advertiserId, advertiserId)) {
-			this.advertiserId = advertiserId;
 			notifyDataChanged();
 		}
 	}